src: more uint_* pruning

+++ src/version.h
@@ -1,2 +1,2 @@
-static const uint8_t citrun_major = 0;
-static const uint8_t citrun_minor = 0;
+static const unsigned int citrun_major = 0;
+static const unsigned int citrun_minor = 0;
